+++ src/test/java/io/zipcoder/casino/PlayerTest/PlayerTest.java
@@ -23,12 +23,12 @@ public class PlayerTest {
     public void constructorWithNameAgeBalanceTest() {
         String expectedName = "Suzy";
         Integer expectedAge = 24;
-        Double expectedBalance = 100.00;
+        Integer expectedBalance = 100;
         Player player1 = new Player(expectedName, expectedAge, expectedBalance);
 
         String actualName = player1.getName();
         Integer actualAge = player1.getAge();
-        Double actualBalance = player1.getBalance();
+        Integer actualBalance = player1.getBalance();
 
         Assert.assertEquals(expectedName, actualName);
         Assert.assertEquals(expectedAge, actualAge);
@@ -76,20 +76,20 @@ public class PlayerTest {
     }
     @Test
     public void setBalanceTest() {
-        Player player1 = new Player("Bob", 49, 100.00);
-        Double expected = 100.00;
+        Player player1 = new Player("Bob", 49, 100);
+        Integer expected = 100;
         player1.setBalance(expected);
-        Double actual = player1.getBalance();
+        Integer actual = player1.getBalance();
         Assert.assertEquals(expected, actual);
 
     }
 
     @Test
     public void getBalanceTest() {
-        Player player1 = new Player("George", 51, 200.00);
-        Double expected = 200.00;
+        Player player1 = new Player("George", 51, 200;
+        Integer expected = 200;
 
-        Double actual = player1.getBalance();
+        Integer actual = player1.getBalance();
         Assert.assertEquals(expected, actual);
 
     }
